Fixed the following issues with inventory: - Inventory SQL table missing 'Type' column - Promises not properly chaining in merchService and app.js - Base image storage path not servable, moved to static/media - Images not returned to app.js in useable format (JSON.parse(binarydata) doesn't work)
Fixed the following issues with app structure:
- npm start was running with --inspect
-- Replaced with `npm run debug`
- Added media folder to gitignore

!!Please note that we should not be sending file data back from the server.
Media can be served from the static folder by file path, I have changed the inventory.js
to use this and add the items to the page.

It's possible we'll need to be more clever about this later, but for now it's fine.